alloy/processors/ai_sorter/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
requests==2.32.3
pydantic==2.9.2
pyyaml==6.0.2
pytest==8.3.3
httpx==0.27.2